Unpick world state circulars.
