+++ yarn-project/world-state/src/synchronizer/world_state_synchronizer.ts
@@ -1,4 +1,4 @@
-import { MerkleTreeOperations } from '../index.js';
+import { MerkleTreeOperations } from '../world-state-db/index.js';
 
 /**
  * Defines the possible states of the world state synchronizer.
