+++ yarn-project/world-state/src/world-state-db/index.ts
@@ -1,2 +1,3 @@
 export * from './merkle_trees.js';
 export * from './merkle_tree_db.js';
+export * from './merkle_tree_operations.js';
